Updated settings.styl to match that of Axis 0.5.x.
This was done by curl'ing the raw settings.styl from Axis itself, I
don't see any issues with this method.

I also updated the link that is pointed to in the header of the file.
This is hopefully a more accurate one, as the previous on 404'd. An
alternative might be the Axis docs themselves, but they seem a little
dated for now.